backend/requirements.txt
aniso8601==9.0.1
asgiref==3.4.1
click==8.0.1
colorama==0.4.4
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
Flask==2.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-pymysql==0.2.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.13.0
PyMySQL==1.0.2
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.2
Werkzeug==2.0.1
